--- src/main/scala/com/activegrid/model/BaseEntity.scala
@@ -1,8 +1,0 @@
-package com.activegrid.model
-
-/**
-  * Created by shareefn on 26/9/16.
-  */
-trait BaseEntity {
-  val id: Option[Long]
-}
