+++ src/main/scala/com/imaginea/Main.scala
@@ -28,7 +28,6 @@ object Main extends App {
   implicit val executionContext = system.dispatcher
   val logger = Logger(LoggerFactory.getLogger(getClass.getName))
 
-
   implicit object KeyPairStatusFormat extends RootJsonFormat[KeyPairStatus] {
     override def write(obj: KeyPairStatus): JsValue = JsString(obj.name.toString)
 
@@ -48,9 +47,21 @@ object Main extends App {
   implicit val softwareFormat = jsonFormat(Software.apply, "id", "version", "name", "provider", "downloadURL", "port", "processNames", "discoverApplications")
   implicit val softwarePageFormat = jsonFormat4(Page[Software])
   implicit val ImageFormat = jsonFormat(ImageInfo.apply, "id", "state", "ownerId", "publicValue", "architecture", "imageType", "platform", "imageOwnerAlias", "name", "description", "rootDeviceType", "rootDeviceName", "version")
-  implicit val PageFormat = jsonFormat4(Page[ImageInfo])
+  implicit val PageImageFormat = jsonFormat4(Page[ImageInfo])
   implicit val appSettingsFormat = jsonFormat(AppSettings.apply, "id", "settings", "authSettings")
-  implicit val siteFormat = jsonFormat(Site.apply, "id", "siteName", "groupBy")
+
+  implicit val portRangeFormat = jsonFormat(PortRange.apply, "id", "fromPort", "toPort")
+  implicit val sshAccessInfoFormat = jsonFormat(SSHAccessInfo.apply, "id", "keyPair", "userName", "port")
+  implicit val instanceConnectionFormat = jsonFormat(InstanceConnection.apply, "id", "sourceNodeId", "targetNodeId", "portRanges")
+  implicit val processInfoFormat = jsonFormat(ProcessInfo.apply, "id", "pid", "ParentPid", "name", "command", "owner", "residentBytes", "software", "softwareVersion")
+  implicit val instanceUserFormat = jsonFormat(InstanceUser.apply, "id", "userName", "publicKeys")
+  implicit val InstanceFlavorFormat = jsonFormat(InstanceFlavor.apply, "name", "cpuCount", "memory", "rootDisk")
+  implicit val PageInstFormat = jsonFormat4(Page[InstanceFlavor])
+  implicit val storageInfoFormat = jsonFormat(StorageInfo.apply, "id", "used", "total")
+  implicit val KeyValueInfoFormat = jsonFormat(KeyValueInfo.apply, "id", "key", "value")
+  implicit val instanceFormat = jsonFormat(Instance.apply, "id", "instanceId", "name", "state", "instanceType", "platform", "architecture", "publicDnsName", "launchTime", "memoryInfo", "rootDiskInfo", "tags", "sshAccessInfo", "liveConnections", "estimatedConnections", "processes", "image", "existingUsers")
+  implicit val PageInstanceFormat = jsonFormat4(Page[Instance])
+  implicit val siteFormat = jsonFormat(Site.apply, "id", "instances", "siteName", "groupBy")
 
   implicit object apmProviderFormat extends RootJsonFormat[APMProvider] {
     override def write(obj: APMProvider): JsValue = {
@@ -565,10 +576,90 @@ object Main extends App {
             complete(StatusCodes.BadRequest, "Unable to Retrieve Softwares List.")
         }
       }
+    } ~ path("instanceTypes" / IntNumber) { siteId =>
+      get {
+        val listOfInstanceFlavors = Future {
+          val mayBeSite = Site.fromNeo4jGraph(siteId)
+          mayBeSite match {
+            case Some(site) =>
+              val listOfInstances = site.instances
+              val listOfInstanceFlavors = listOfInstances.map(instance => InstanceFlavor(instance.instanceType.get, None, instance.memoryInfo.get.total, instance.rootDiskInfo.get.total))
+              Page[InstanceFlavor](listOfInstanceFlavors)
+
+            case None =>
+              logger.warn(s"Failed while doing fromNeo4jGraph of Site for siteId : $siteId")
+              Page[InstanceFlavor](List.empty[InstanceFlavor])
+          }
+        }
+        onComplete(listOfInstanceFlavors) {
+          case Success(successResponse) => complete(StatusCodes.OK, successResponse)
+          case Failure(ex) =>
+            logger.error(s"Unable to get List; Failed with ${ex.getMessage}", ex)
+            complete(StatusCodes.BadRequest, "Unable to get List of Instance Flavors")
+        }
+      }
     }
   }
 
-  val route: Route = userRoute ~ keyPairRoute ~ catalogRoutes ~ appSettingServiceRoutes ~ apmServiceRoutes
+  def nodeRoutes = pathPrefix("node") {
+    path("list") {
+      get {
+        val listOfAllInstanceNodes = Future {
+          logger.info("Received GET request for all nodes")
+          val label: String = "Instance"
+          val nodesList = GraphDBExecutor.getNodesByLabel(label)
+          val instanceList = nodesList.flatMap(node => Instance.fromNeo4jGraph(node.getId))
+          Page[Instance](instanceList)
+        }
+        onComplete(listOfAllInstanceNodes) {
+          case Success(successResponse) => complete(StatusCodes.OK, successResponse)
+          case Failure(ex) =>
+            logger.error(s"Unable to get Instance nodes; Failed with ${ex.getMessage}", ex)
+            complete(StatusCodes.BadRequest, s"Unable to get Instance nodes")
+        }
+      }
+    } ~ path("topology") {
+      get {
+        val topology = Future {
+          logger.debug("received GET request for topology")
+          Page[Instance](List.empty[Instance])
+        }
+        onComplete(topology) {
+          case Success(successResponse) => complete(StatusCodes.OK, successResponse)
+          case Failure(ex) =>
+            logger.error(s"Unable to get Instance nodes; Failed with ${ex.getMessage}", ex)
+            complete(StatusCodes.BadRequest, s"Unable to get Instance nodes")
+        }
+      }
+    } ~ path(Segment) { name =>
+      get {
+        val nodeInstance = Future {
+          logger.info(s"Received GET request for node - $name")
+          if (name == "localhost") {
+            val instance = Instance(name)
+            instance.toNeo4jGraph(instance)
+          }
+          val instanceNode = GraphDBExecutor.getNodeByProperty("Instance", "name", name)
+          instanceNode match {
+            case Some(node) => Instance.fromNeo4jGraph(node.getId).get
+            case None =>
+              val name = "echo node"
+              val tags: List[KeyValueInfo] = List(KeyValueInfo(None, "tag", "tag"))
+              val processInfo = ProcessInfo(1, 1, "init")
+              Instance(name, tags, Set(processInfo))
+          }
+        }
+        onComplete(nodeInstance) {
+          case Success(successResponse) => complete(StatusCodes.OK, successResponse)
+          case Failure(ex) =>
+            logger.error(s"Unable to get Instance with name $name; Failed with ${ex.getMessage}", ex)
+            complete(StatusCodes.BadRequest, s"Unable to get Instance with name $name")
+        }
+      }
+    }
+  }
+
+  val route: Route = userRoute ~ keyPairRoute ~ catalogRoutes ~ appSettingServiceRoutes ~ apmServiceRoutes ~ nodeRoutes
 
   val bindingFuture = Http().bindAndHandle(route, config.getString("http.host"), config.getInt("http.port"))
   logger.info(s"Server online at http://${config.getString("http.host")}:${config.getInt("http.port")}")
