+++ src/main/scala/com/imaginea/activegrid/core/models/BaseEntity.scala
@@ -1,7 +1,7 @@
 package com.imaginea.activegrid.core.models
 
 /**
-  * Created by nagulmeeras on 30/09/16.
+  * Created by shareefn on 26/9/16.
   */
 trait BaseEntity {
   val id: Option[Long]
