+++ src/main/scala/com/imaginea/activegrid/core/models/InstanceConnection.scala
@@ -1,4 +1,4 @@
-package com.activegrid.model
+package com.imaginea.activegrid.core.models
 
 import com.typesafe.scalalogging.Logger
 import org.neo4j.graphdb.Node
@@ -53,4 +53,5 @@ object InstanceConnection {
       InstanceConnection.fromNeo4jGraph(id)
     }
   }
+
 }
