+++ src/main/scala/com/imaginea/activegrid/core/models/InstanceFlavor.scala
@@ -1,4 +1,4 @@
-package com.activegrid.model
+package com.imaginea.activegrid.core.models
 
 /**
   * Created by shareefn on 27/9/16.
