+++ src/main/scala/com/imaginea/activegrid/core/models/KeyValueInfo.scala
@@ -1,4 +1,4 @@
-package com.activegrid.model
+package com.imaginea.activegrid.core.models
 
 import com.typesafe.scalalogging.Logger
 import org.neo4j.graphdb.Node
@@ -42,4 +42,5 @@ object KeyValueInfo {
       KeyValueInfo.fromNeo4jGraph(id)
     }
   }
+
 }
