+++ src/main/scala/com/imaginea/activegrid/core/models/PortRange.scala
@@ -1,4 +1,4 @@
-package com.activegrid.model
+package com.imaginea.activegrid.core.models
 
 import com.typesafe.scalalogging.Logger
 import org.neo4j.graphdb.Node
@@ -41,4 +41,5 @@ object PortRange {
       PortRange.fromNeo4jGraph(id)
     }
   }
+
 }
